+++ apps/tenant-management-webapp-e2e/src/integration/common/common.page.ts
@@ -92,5 +92,55 @@ class Common {
   loadMoreButton() {
     return cy.xpath('//goa-button[text()="Load more"]');
   }
+
+  //Form app page objects:
+  formAppLoginButton() {
+    return cy.get('#kc-login');
+  }
+
+  formAppUsernameEmailField() {
+    return cy.get('[name=username]');
+  }
+
+  formAppPasswordField() {
+    return cy.get('[name=password]');
+  }
+
+  formAppFormTextField(label) {
+    return cy.xpath(`//goa-form-item[@label="${label}"]/goa-input[@type="text"]`);
+  }
+
+  formAppFormSubmitButton() {
+    return cy.xpath('//goa-button[text()="Submit"]');
+  }
+
+  //Status app page objects
+  statusNotificationPleaseContact() {
+    return cy.xpath('//*[@class="section-vs"]/a');
+  }
+
+  statusNotificationPageTitle() {
+    return cy.get('h2[data-testid="service-name"]');
+  }
+
+  statusNotificationSignupDescription() {
+    return cy.xpath('//h3[text()="Sign up for notifications"]/following-sibling::div/a');
+  }
+
+  applicationStatus(appName) {
+    return cy.xpath(
+      `//b[contains(text(),"${appName}")]/parent::div/following-sibling::div/div[@class="status-button"]/div`
+    );
+  }
+
+  applicationStatusUpdatedTimestamp(appName) {
+    return cy.xpath(
+      `//b[contains(text(),"${appName}")]/ancestor::div[@class="flex-column"]//*[@data-testid="service-created-date"]`
+    );
+  }
+
+  applicationNames() {
+    return cy.xpath('//div[@class="title-line"]/following-sibling::div/div//b');
+  }
 }
 export default Common;
