--- apps/tenant-management-webapp-e2e/src/integration/form-app.feature
@@ -1,2 +1,0 @@
-@form-app
-Feature: Form-app
