+++ apps/tenant-management-webapp-e2e/src/integration/service-status/service-status.step.ts
@@ -421,83 +421,6 @@ function searchNoticeCards(mode, desc, app, startDateTime, endDateTime) {
   });
 }
 
-Given('a tenant admin user is on status applications page', function () {
-  commonlib.tenantAdminDirectURLLogin(
-    Cypress.config().baseUrl,
-    Cypress.env('realm'),
-    Cypress.env('email'),
-    Cypress.env('password')
-  );
-  commonlib.tenantAdminMenuItem('Status', 4000);
-  commonObj.serviceTab('Status', 'Applications').click();
-  cy.wait(2000); // Applications page is slow to load applications and healt check info
-});
-
-When('the user {string} the subscribe checkbox for health check notification type', function (checkboxOperation) {
-  statusObj
-    .applicationHealthChangeNotificationSubscribeCheckbox()
-    .shadow()
-    .find('.goa-checkbox-container')
-    .invoke('attr', 'class')
-    .then((classAttVal) => {
-      if (classAttVal == undefined) {
-        expect.fail('Failed to get subscribe checkbox class attribute value.');
-      } else {
-        switch (checkboxOperation) {
-          case 'selects':
-            if (classAttVal.includes('selected')) {
-              cy.log('The subscribe checkbox was already checked.');
-            } else {
-              statusObj
-                .applicationHealthChangeNotificationSubscribeCheckbox()
-                .shadow()
-                .find('.goa-checkbox-container')
-                .click();
-            }
-            break;
-          case 'unselects':
-            if (classAttVal.includes('selected')) {
-              statusObj
-                .applicationHealthChangeNotificationSubscribeCheckbox()
-                .shadow()
-                .find('.goa-checkbox-container')
-                .click();
-            } else {
-              cy.log('The subscribe checkbox was already unchecked.');
-            }
-            break;
-          default:
-            expect(checkboxOperation).to.be.oneOf(['selects', 'unselects']);
-        }
-      }
-    });
-});
-
-Then('the user views the subscribe checkbox is {string}', function (checkboxStatus) {
-  cy.wait(2000); // Wait for the checkbox status to show
-  statusObj
-    .applicationHealthChangeNotificationSubscribeCheckbox()
-    .shadow()
-    .find('.goa-checkbox-container')
-    .invoke('attr', 'class')
-    .then((classAttVal) => {
-      if (classAttVal == undefined) {
-        expect.fail('Failed to get subscribe checkbox class attribute value.');
-      } else {
-        switch (checkboxStatus) {
-          case 'checked':
-            expect(classAttVal).to.contain('selected');
-            break;
-          case 'unchecked':
-            expect(classAttVal).to.not.contain('selected');
-            break;
-          default:
-            expect(checkboxStatus).to.be.oneOf(['checked', 'unchecked']);
-        }
-      }
-    });
-});
-
 When('the user clicks Add application button', function () {
   statusObj.addApplicationButton().shadow().find('button').click({ force: true });
   cy.wait(1000); // Add a wait to avoid accessibility test to run too quickly before the modal is fully loaded
