--- apps/tenant-management-webapp-e2e/src/integration/status-app.feature
@@ -1,2 +1,0 @@
-@status-app
-Feature: Status-app
